Featured: Adding support for featured conventional commits
Inspired by a random thought I had on mastodon: this site now generates a `/changelog` page that is generated from the `git` commit history. Aside from general statistics. it will grab the *contents* of commits like this one and render them as *Featured Changesets* during compilation. I don't know if I've ever seen anybody else talk about doing something _quite_ like this, so I think this could be pretty *interesting* to explore in the future.

I'm a big fan of `git blame` and other VCS-centric annotation systems. I have always *tried* to curate *meaningful* commit messages as the "ideal". So, being able to embed a microblog entry in the commit history, but still have them accessible on the web feels really nice. I also think that it might, at least for me, lower the barrier to entry to *write about* the changes that I make to the site, that maybe I normally *wouldn't*. Because while I don't *have* to write a blog about my work, I *always* make a commit.

So, we'll see where this goes… but I'm _excited_.